Move encoding utils to identity_core
