+++ identity_core/src/utils/has_id.rs
@@ -1,0 +1,7 @@
+use core::hash::Hash;
+
+pub trait HasId {
+    type Id: Hash + PartialEq + Eq + PartialOrd + Ord;
+
+    fn id(&self) -> &Self::Id;
+}
