+++ identity_proof/src/error.rs
@@ -1,9 +1,5 @@
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
-    #[error("Failed to decode base58 data: {0}")]
-    DecodeBase58(#[from] bs58::decode::Error),
-    #[error("Failed to decode base64 data: {0}")]
-    DecodeBase64(#[from] base64::DecodeError),
     #[error("Failed to canonicalize object: {0}")]
     Canonicalize(anyhow::Error),
     #[error("Failed to pre-process document: {0}")]
