--- identity_proof/src/utils/mod.rs
@@ -1,3 +1,0 @@
-mod encoding;
-
-pub use self::encoding::*;
